backend/requirements.txt
blinker==1.9.0
click==8.3.2
colorama==0.4.6
exceptiongroup==1.3.1
Flask==3.1.3
flask-cors==6.0.2
Flask-SQLAlchemy==3.1.1
greenlet==3.4.0
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.1
pluggy==1.6.0
Pygments==2.20.0
pytest==9.0.3
python-dotenv==1.2.2
SQLAlchemy==2.0.49
tomli==2.4.1
typing_extensions==4.15.0
Werkzeug==3.1.8
openai==1.59.8
google-genai==1.73.1

